@@ -1,13 +1,40 @@
-[project]
-name = 'SQLDbWrpr'
-version = '4.1.4'
-authors = [
-    { name = 'Hendrik du Toit', email = 'hendrik@brightedge.co.za' },
-    { name = 'Henru du Toit', email = 'henru@brightedge.co.za' },
-    { name = 'Dirk du Toit', email = 'dirk@brightedge.co.za' },
+[build-system]
+requires = [
+    "poetry-core>=1.0.0",
 ]
+build-backend = "poetry.core.masonry.api"
+
+[tool.black]
+line-length = 120
+#skip-string-normalization = true
+target-version = [
+    "py312",
+]
+
+[tool.codespell]
+count = ''
+quiet-level = 2
+skip = 'documentation/Archive/*'
+ignore-words-list = 'space-holder'
+
+[tool.coverage.run]
+source = [
+    'src',
+    'tests',
+]
+
+[tool.isort]
+profile = "black"
+
+[tool.poetry]
+name = 'SQLDbWrpr'
+version = '4.2.0'
 description = 'Wrapper for importing CSV and Text files into MySQL and Postgress'
-license = {file = 'LICENSE'}
+authors = [
+    "Hendrik du Toit <hendrik@brightedge.co.za>",
+]
+license = "MIT"
+readme = "README.rst"
 classifiers = [
     'Development Status :: 5 - Production/Stable',
     'Intended Audience :: Developers',
@@ -18,66 +45,50 @@ classifiers = [
     'License :: OSI Approved :: MIT License',
     'Programming Language :: Python :: 3.10',
     'Programming Language :: Python :: 3.11',
+    'Programming Language :: Python :: 3.12',
 ]
-readme = {file = 'README.rst', content-type = 'text/x-rst'}
-requires-python = '>=3.8, <=12.0'
-dependencies = [
-    'beetools',
-    'csvwrpr',
-    'displayfx',
-    'fixdate',
-    'pyodbc',
-    'mysql.connector'
-  ]
+# documentation = "https://readthedocs.org"
+homepage = 'https://github.com/BrightEdgeeServices/SQLDbWrpr'
+repository = "https://github.com/RealTimeEvents/SQLDbWrpr.git"
 
-[project.optional-dependencies]
-dev = [
-    'git-it',
-    'codecov',
-    'docutils',
-    'pytest',
-    'pytest-cov',
-    'rstcheck',
-]
+[tool.poetry.dependencies]
+python = "^3.12.2"
+mysql-connector-python = "^9.1.0"
+csvwrpr = "^1.16.0"
+fixdate = "^2.0.0"
+pyodbc = "^5.2.0"
 
-[project.urls]
-Changelog = 'https://github.com/RealTimeEvents/rte-api/blob/enhancement/master/CHANGES.md'
-Homepage = 'https://github.com/RealTimeEvents/rte-api'
+[tool.poetry.group.dev.dependencies]
+mdformat-gfm = "^0.3.6"
+mdformat-frontmatter = "^2.0.8"
+mdformat-footnote = "^0.1.1"
+black = "^24.8.0"
+flake8 = "^7.1.1"
+pre-commit = "^3.8.0"
+pytest = "^8.3.4"
+codecov = "^2.1.13"
+pytest-cov = "^6.0.0"
+rstcheck = "^6.2.4"
+pydantic-settings = "^2.7.1"
 
-[build-system]
-requires = [
-    "setuptools",
-    "wheel",
-    "setuptools-git >= 1.1",
-#    "pytest-runner",
-#    "toml",
-#    "reahl-component-metadata>=6.1,<6.2'",
-#    "setuptools_scm"
-]
-build-backend = "setuptools.build_meta"
-
-[tool.black]
-skip-string-normalization = true
-line-length = 120
-target-version = ["py311"]
-
-[tool.codespell]
-count = ''
-quiet-level = 2
-skip = 'documentation/Archive/*'
-
-[tool.coverage.run]
-source = ['src']
+[tool.poetry.urls]
+issues = "https://github.com/BrightEdgeeServices/SQLDbWrpr/issues"
+changelog = 'https://github.com/BrightEdgeeServices/SQLDbWrpr/blob/enhancement/master/ReleaseNotes.md'
 
 [tool.pytest.ini_options]
-addopts = ['--import-mode=importlib', '-vv', '--ignore-glob=*\\Archive']
+addopts = [
+    '-vv',
+    '--ignore-glob=*\\Archive',
+]
 #addopts = ['--import-mode=importlib', "-v", "--cov=./", "--ignore-glob=*\\Archive", "--cov-report=html"]
-pythonpath = ['src','src\resultcollector', 'tests' ]
 markers = [
     "select: Run a selection of tests",
 ]
-#filterwarnings = [
-#    "ignore::DeprecationWarning",
-#]
-#python_functions = 'demo_setup'
+filterwarnings = [
+    "ignore::DeprecationWarning",
+]
+pythonpath = [
+    'src',
+    'tests',
+]
 testpaths = 'tests'
